requirements.txt
llama-index==0.10.0
llama-index-llms-azure-openai>=0.1.0
llama-index-embeddings-azure-openai>=0.1.0
gradio>=4.0.0
python-dotenv>=0.21.0
openai>=1.0.0